wsd: move process wrappers into their own home
