+++ wsd/COOLWSD.hpp
@@ -34,6 +34,8 @@
 #include <Poco/Util/OptionSet.h>
 #include <Poco/Util/ServerApplication.h>
 
+class WSProcess;
+class ForKitProcess;
 class ChildProcess;
 class TraceFileWriter;
 class DocumentBroker;
@@ -41,198 +43,6 @@ class ClipboardCache;
 class FileServerRequestHandler;
 
 std::shared_ptr<ChildProcess> getNewChild_Blocks(SocketPoll &destPoll, unsigned mobileAppDocId);
-
-// A WSProcess object in the WSD process represents a descendant process, either the direct child
-// process ForKit or a grandchild Kit process, with which the WSD process communicates through a
-// WebSocket.
-class WSProcess
-{
-public:
-    /// @param pid is the process ID.
-    /// @param socket is the underlying Socket to the process.
-    WSProcess(const std::string& name,
-              const pid_t pid,
-              const std::shared_ptr<StreamSocket>& socket,
-              std::shared_ptr<WebSocketHandler> handler) :
-
-        _name(name),
-        _pid(pid),
-        _ws(std::move(handler)),
-        _socket(socket)
-    {
-        LOG_INF(_name << " ctor [" << _pid << "].");
-    }
-
-    WSProcess(WSProcess&& other) = delete;
-
-    const WSProcess& operator=(WSProcess&& other) = delete;
-
-    virtual ~WSProcess()
-    {
-        LOG_DBG('~' << _name << " dtor [" << _pid << "].");
-
-        if (_pid <= 0)
-            return;
-
-        terminate();
-
-        // No need for the socket anymore.
-        _ws.reset();
-        _socket.reset();
-    }
-
-    /// Let the child close a nice way.
-    void close()
-    {
-        if (_pid < 0)
-            return;
-
-        try
-        {
-            LOG_DBG("Closing ChildProcess [" << _pid << "].");
-
-            requestTermination();
-
-            // Shutdown the socket.
-            if (_ws)
-                _ws->shutdown();
-        }
-        catch (const std::exception& ex)
-        {
-            LOG_ERR("Error while closing child process: " << ex.what());
-        }
-
-        _pid = -1; // Detach from child.
-    }
-
-    /// Request graceful termination.
-    void requestTermination()
-    {
-        // Request the child to exit
-        if (isAlive())
-        {
-            LOG_DBG("Stopping ChildProcess [" << _pid << "] by sending 'exit' command");
-            sendTextFrame("exit", /*flush=*/true);
-        }
-    }
-
-    /// Kill or abandon the child.
-    void terminate()
-    {
-        if (_pid < 0)
-            return;
-
-#if !MOBILEAPP
-        if (::kill(_pid, 0) == 0)
-        {
-            LOG_INF("Killing child [" << _pid << "].");
-#if CODE_COVERAGE || VALGRIND_COOLFORKIT
-            constexpr auto signal = SIGTERM;
-#else
-            constexpr auto signal = SIGKILL;
-#endif
-            if (!SigUtil::killChild(_pid, signal))
-            {
-                LOG_ERR("Cannot terminate lokit [" << _pid << "]. Abandoning.");
-            }
-        }
-#else
-        // What to do? Throw some unique exception that the outermost call in the thread catches and
-        // exits from the thread?
-#endif
-        _pid = -1;
-    }
-
-    pid_t getPid() const { return _pid; }
-
-    /// Send a text payload to the child-process WS.
-    bool sendTextFrame(const std::string& data, bool flush = false)
-    {
-        return sendFrame(data, false, flush);
-    }
-
-    /// Send a payload to the child-process WS.
-    bool sendFrame(const std::string& data, bool binary = false, bool flush = false)
-    {
-        try
-        {
-            if (_ws)
-            {
-                LOG_TRC("Send to " << _name << " message: ["
-                                   << COOLProtocol::getAbbreviatedMessage(data) << ']');
-                _ws->sendMessage(data.c_str(), data.size(),
-                                 (binary ? WSOpCode::Binary : WSOpCode::Text), flush);
-                return true;
-            }
-        }
-        catch (const std::exception& exc)
-        {
-            LOG_ERR("Failed to send " << _name << " [" << _pid << "] data [" <<
-                    COOLProtocol::getAbbreviatedMessage(data) << "] due to: " << exc.what());
-            throw;
-        }
-
-        LOG_WRN("No socket to " << _name << " to send [" << COOLProtocol::getAbbreviatedMessage(data) << ']');
-        return false;
-    }
-
-    /// Check whether this child is alive and socket not in error.
-    /// Note: zombies will show as alive, and sockets have waiting
-    /// time after the other end-point closes. So this isn't accurate.
-    virtual bool isAlive() const
-    {
-#if !MOBILEAPP
-        try
-        {
-            return _pid > 1 && _ws && ::kill(_pid, 0) == 0;
-        }
-        catch (const std::exception&)
-        {
-        }
-
-        return false;
-#else
-        return _pid > 1;
-#endif
-    }
-
-protected:
-    std::shared_ptr<WebSocketHandler> getWSHandler() const { return _ws; }
-    std::shared_ptr<StreamSocket> getSocket() const { return _socket.lock(); };
-
-private:
-    std::string _name;
-    std::atomic<pid_t> _pid; ///< The process-id, which can be access from different threads.
-    std::shared_ptr<WebSocketHandler> _ws;  // FIXME: should be weak ? ...
-    std::weak_ptr<StreamSocket> _socket;
-};
-
-#if !MOBILEAPP
-
-class ForKitProcWSHandler final : public WebSocketHandler
-{
-public:
-    template <typename T>
-    ForKitProcWSHandler(const std::weak_ptr<StreamSocket>& socket, const T& request)
-        : WebSocketHandler(socket.lock(), request)
-    {
-    }
-
-    virtual void handleMessage(const std::vector<char>& data) override;
-};
-
-class ForKitProcess final : public WSProcess
-{
-public:
-    template <typename T>
-    ForKitProcess(int pid, std::shared_ptr<StreamSocket>& socket, const T& request)
-        : WSProcess("ForKit", pid, socket, std::make_shared<ForKitProcWSHandler>(socket, request))
-    {
-        socket->setHandler(getWSHandler());
-    }
-};
-
-#endif
 
 /// The Server class which is responsible for all
 /// external interactions.
